# Remove dependency on `equator` to speed up compile times