--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,7 +10,6 @@ license = "MIT"
 keywords = ["vec", "allocation", "box", "slice", "alignment"]
 
 [dependencies]
-equator = "0.4.1"
 serde = { version = "1.0", optional = true, default-features = false }
 
 [dev-dependencies]