--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1395,7 +1395,7 @@ mod serde_tests {
     #[test]
     fn can_limit_deserialization_size() {
         // Malformed serialized data indicating a sequence of length u64::MAX.
-        let ser = vec![
+        let ser = alloc::vec![
             253, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 1, 1, 1, 1, 1, 1, 1, 1,
         ];