@sarah-quinones 2025-09-20
the crate uses `equator::assert` for pretty printing of assertions, so i don't wanna get rid of it for now